Minor improvements
If the crash is caused by a breakpoint instruction the address of the
breakpoint is now displayed as the faulting instruction.
If the process crashed because it tried to execute invalid memory (fa ==
pc) then it is not possible to determine the region the crash happend in
and the relative offset to it's base address. These values are now
omitted.